i hope im doing this right
added my code that i never added mb and also tried to make it so that lukes code is visualized in pygame but idk if its 100% correct. its working when i test it at least. also i did mine in a seperate file just in case i messed something up when experimenting. ill work on the score over the weekend, im in oregon rn.